Added quick fix to undocking sequence. The issue was that the XRGrabInteractable caches the isKinematic and useGravity state on entered, sets isKinemtic true and useGraivy false while being held, and finally reinstates the caches settings on exited.
When a gun is docked, it's state is set to isKinematic true useGravity false (like the interactable), but when the interactable caches it prior to calling the local entered callback method, it stores it in the wrong state. When picked up from the dock (entered) it reinstates the wrong settings, but the callback overrides those settings from it's own cache. This however does not carry forward as when next released (exited)
it incorrectly applys the wrong settings again which prevents gravity from kicking in.

The hacky solution was to cache the interactable properties upon awake and simply override the rigid body OnSelectExited.